Introduce NodeVisitor class
