+++ composer.json
@@ -9,7 +9,8 @@
   },
   "require-dev": {
     "phpunit/phpunit": "^10.5",
-    "squizlabs/php_codesniffer": "^3.12"
+    "squizlabs/php_codesniffer": "^3.12",
+    "phpstan/phpstan": "^2.1"
   },
   "autoload": {
     "psr-4": {
